% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_commstats.r
\name{calc_commstats}
\alias{calc_commstats}
\title{Calculate community statistics}
\usage{
calc_commstats(comm, topo_names, pools)
}
\arguments{
\item{comm}{(required) site x microsite matrix with integers representing the 
associations present in each microsite}

\item{topo_names}{(required) host x symbiont association matrix with integers
identifying each association (as returned by \code{name_topo})}

\item{pools}{(required) named list of pools of hosts as symbionts present at
each site. Pools are generated from the community matrix by 
\code{calc_pool}. The host pool should be named \code{'a'} and the symbiont 
pool named \code{'b'}.}
}
\value{
a dataframe of community statistics (see details) with rows 
	representing sites
}
\description{
Calculate diversity and abundance of hosts, symbionts and associations
}
\details{
This function calculates abundance and diversity of hosts, symbionts
and associations that are established at sites. If a species is present
in a site's species pool, but not in an association at a microsite it 
will not be counted. Three diversity statistics are calculated for hosts
(\code{'a'}), symbionts (\code{'b'}) and associations (\code{'species'}):
\itemize{
\item \strong{\code{S}} : species richness (number of species present
	at each site), see \code{\link{calc_rich}}
\item \strong{\code{Stot}} : total species richness across all sites,
	see \code{\link{calc_tot_rich}}
\item \strong{\code{Beta}} : beta diversity (total richness divided
	by mean richness at each site)
}
The function also returns the number of microsites occupied at each site
(\code{N}), the total number of microsites occupied across the entire
metacommunity (\code{Ntot}), and the pearson correlation between 
species richness of hosts and symbionts across sites (\code{Core_ab}).
Community statistics are returned as a dataframe with each row representing
a site.
}

