% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_diss.r
\name{calc_diss}
\alias{calc_diss}
\title{Calculate community dissimilarity}
\usage{
calc_diss(comm, topo_names, type, metric, binary = F, pool = NULL)
}
\arguments{
\item{comm}{(required) site x microsite matrix with integers representing
the associations present in each microsite}

\item{topo_names}{(required) host x symbiont association matrix with 
integers identifying each association (as returned by \code{name_topo})}

\item{type}{(required) character string indicating whether the function
should do calculations on host species (\code{'a'}), symbiont
species (\code{'b'}), or associations (\code{'species'})}

\item{metric}{(required) dissimilarity metric used for calculation
(see \code{\link{vegdist}}}

\item{binary}{logical indicating whether dissimilarties should be calculated
from species abundances (\code{FALSE}, default) or presence/absence
(\code{TRUE})}

\item{pool}{site x species matrix for the species pool present at each
site (as returned by \code{\link{calc_pool}}). Must be supplied if 
calculating for host or symbiont species.}
}
\value{
matrix of site dissimilarities, as returned by 
	\code{\link[vegan]{vegdist}}
}
\description{
Calculate dissimilarity in species composition among sites
}
\details{
This function uses \code{\link[vegan]{vegdist}} in the vegan package to
calculate community dissimilarity among sites. Dissimilarity can be
calculated for associations or for host or symbiont species present
in established associations. Any metric implemented by 
\code{\link[vegan]{vegdist}} can be used and dissimilarities can be 
calculated based on abundances or based on presence/absence 
(\code{binary=TRUE}).
}

