% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_envcorr.r
\name{calc_envcorr}
\alias{calc_envcorr}
\title{Community-environment correlations}
\usage{
calc_envcorr(comm, topo_names, env, metric = NULL, binary, pools = NULL)
}
\arguments{
\item{comm}{(required) site x microsite matrix with integers representing the 
associations present in each microsite}

\item{topo_names}{(required) host x symbiont association matrix with integers
identifying each association (as returned by \code{name_topo})}

\item{env}{(required) site x env matrix with environmental conditions at 
each site}

\item{metric}{vector of dissimilarity metric (see 
\code{\link[vegan]{vegdist}} for options). Defaults to \code{NULL} for none.}

\item{binary}{lvector of logicals indicating whether calculations should be
performed on species abundances (\code{FALSE}, default) or presence/absence
(\code{TRUE}) or both}

\item{pools}{(required) named list of pools of hosts as symbionts present at
each site. Pools are generated from the community matrix by 
\code{calc_pool}. The host pool should be named \code{'a'} and the symbiont 
pool named \code{'b'}.}
}
\value{
array with dimensions \code{[community type, environmental axis, 
	measurement, abundances vs presence/absence]} (see details)
}
\description{
Calculates multiple community-environment correlations
}
\details{
This summary function calculates the correlation between each
environmental axis (provided in \code{env}) and species richness, abundance
and community structure for established associations and host and symbiont
species involved in established associations. For each environmental axis and 
dissimilarity metric given in \code{metric} it calculates the correlation 
between a community dissimilarity matrix (see \code{link{calc_diss}}) and
the environmental distance matrix. Finally, the function conducts a 
constrained ordination along each environmental axis (see  
\code{link{calc_rda}}) and returns the adjusted \eqn{R^2}. These statistics
are returned as a 4-dimensional array where the first dimension indicates 
the community examined (associations: \code{'species'}, hosts: \code{'a'}, 
symbionts: \code{'b'}), the second dimension indicates which environmental
axis was used, the third dimension indicates the statistic (richness:
\code{'S'}, abundance: \code{'N'}, RDA: \code{'rda'}, or correlation with
community dissimilarity metric), and the fourth dimension indicates whether
the calculations are based on abundances (\code{FALSE} or presence/absence
(\code{TRUE}).
}

