% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_network.r
\name{calc_network}
\alias{calc_network}
\title{Calculate association network}
\usage{
calc_network(comm, topo_names, bysite = F)
}
\arguments{
\item{comm}{(required) site x microsite matrix with integers representing
the associations present in each microsite}

\item{topo_names}{(required) host x symbiont association matrix with 
integers identifying each association (as returned by \code{name_topo})}

\item{bysite}{logical indicating whether networks should be calculated for
each site in \code{comm}}
}
\value{
a binary matrix or array representing the association network(s)
}
\description{
Calculates associations network from observed established communities
}
\details{
This function calculates the observed species association network from 
a metacommunity (\code{comm}). It can return either a single matrix
across the entire metacommunity or an array of matrices, one for each 
site, if \code{bysite=TRUE}.
}

