% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_occ.r
\name{calc_occ}
\alias{calc_occ}
\alias{calc_tot_occ}
\title{Calculate abundance}
\usage{
calc_occ(comm)

calc_tot_occ(comm)
}
\arguments{
\item{comm}{(required) site x microsite matrix with integers representing
the associations present in each microsite}
}
\value{
vector of number of occupied microsites at each site or the total
	number of microsites occupied in the entire metacommunity
}
\description{
Calculates the number of occupied microsites at each site or across the 
entire metacommunity
}
\section{Functions}{
\itemize{
\item \code{calc_occ}: Calculate abundance at each site

\item \code{calc_tot_occ}: Calculate total abundance for the metacommunity
}}

