% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pool.r
\name{calc_pool}
\alias{calc_pool}
\title{Calculate species pool}
\usage{
calc_pool(comm, topo_names, partner)
}
\arguments{
\item{comm}{(required) metacommunity matrix indicating which 
association is present in microsite of each site. Rows are sites
and columns are microsites/individuals.}

\item{topo_names}{(required) matrix representing association network 
where integers label each association, as produced by 
\code{\link{name_topo}}. Labels should match those used in \code{comm}.}

\item{partner}{(required) integer indicating whether the function should
calculate the species pool for hosts (\code{1}) or symbionts (\code{2}).}
}
\value{
site x species matrix indicating species presence in each site
}
\description{
Determines which species are present in a community.
}

