% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rda.r
\name{calc_rda}
\alias{calc_rda}
\title{Community-environment RDA}
\usage{
calc_rda(comm, topo_names, type, env, binary = F, pool = NULL)
}
\arguments{
\item{comm}{(required) site x microsite matrix with integers representing
the associations present in each microsite}

\item{topo_names}{(required) host x symbiont association matrix with 
integers identifying each association (as returned by \code{name_topo})}

\item{type}{(required) character string indicating whether the function
should do calculations on host species (\code{'a'}), symbiont
species (\code{'b'}), or associations (\code{'species'})}

\item{env}{(required) matrix of environmental variables across sites}

\item{binary}{logical indicating whether rda should be performed on
species abundances (\code{FALSE}, default) or presence/absence
(\code{TRUE})}

\item{pool}{site x species matrix for the species pool present at each
site (as returned by \code{\link{calc_pool}}). Must be supplied if 
calculating for host or symbiont species.}
}
\description{
Calculate community ~ environment RDA
}
\details{
The function calculates the correspondence between community composition
and a given environmental data matrix (\code{env}) for established
associations or host or symbiont species in established associations
by performing a constrained ordination (\code{\link[vegan]{rda}}) on 
Hellinger-transformed species abundances (as implemented 
by \code{\link[vegan]{decostand}}).
The function returns the adjusted eqn{R^2} as a measured of correspondence
between community structure and environment.
}
\seealso{
\code{\link[vegan]{rda}} for constrained ordination
}

