% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_rich.r
\name{calc_rich}
\alias{calc_rich}
\alias{calc_tot_rich}
\title{Calculate species richness}
\usage{
calc_rich(comm, topo_names, type, pool = NULL)

calc_tot_rich(comm, topo_names, type, pool = NULL)
}
\arguments{
\item{comm}{(required) site x microsite matrix with integers representing
the associations present in each microsite}

\item{topo_names}{(required) host x symbiont association matrix with 
integers identifying each association (as returned by \code{name_topo})}

\item{type}{(required) character string indicating whether the function
should do calculations on host species (\code{'a'}), symbiont
species (\code{'b'}), or associations (\code{'species'})}

\item{pool}{site x species matrix for the species pool present at each
site (as returned by \code{\link{calc_pool}}). Must be supplied if 
calculating for host or symbiont species.}
}
\value{
vector of species richness at each site or the total number of
	species in the entire metacommunity
}
\description{
Calculates species richness at each site or across the entire metacommunity
}
\details{
Calculates the number of host (\code{type = 'a'}) or symbiont 
(\code{type = 'b'}) species present in established associations at each site
or the number of associations (\code{type = 'species'}) at each site.
}
\section{Functions}{
\itemize{
\item \code{calc_rich}: Calculate site richness

\item \code{calc_tot_rich}: Calculate total richness
}}

