% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sa.r
\name{calc_sa}
\alias{calc_sa}
\title{Calculate species abundances}
\usage{
calc_sa(comm, topo_names, type = "species", pool = NULL)
}
\arguments{
\item{comm}{(required) site x microsite matrix with integers representing
the associations present in each microsite}

\item{topo_names}{(required) host x symbiont association matrix with 
integers identifying each association (as returned by \code{name_topo})}

\item{type}{(required) character string indicating whether the function
should do calculations on host species (\code{'a'}), symbiont
species (\code{'b'}), or associations (\code{'species'})}

\item{pool}{site x species matrix for the species pool present at each
site (as returned by \code{\link{calc_pool}}). Must be supplied if 
calculating for host or symbiont species.}
}
\value{
site x species matrix giving the abundace of each species at each
site
}
\description{
Calculate species abundances from a community matrix
}
\details{
Calculates the abundance of associations or host or symbiont
species at each site. The function only counts species
present in associations established at microsites and does
not count species that are only present in a species' pools.
}

