% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_networks.r
\name{compare_networks}
\alias{compare_networks}
\title{Compare association networks}
\usage{
compare_networks(netA, netB)
}
\arguments{
\item{netA}{(required) binary species association matrix}

\item{netB}{(required) binary species association matrix}
}
\value{
a named vector with the number of shared associations 
	(\code{'both'}) and the number of associations unique to each 
	network (\code{'A'} and \code{'B'})
}
\description{
Compares the assocations shared between two association matrices
}
\details{
The function calculates the number of associations shared between two 
association matrices and the number of associations unique to each.
}

