% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/die.r
\name{die}
\alias{die}
\title{Death}
\usage{
die(comm, topo_names, pool, mortality, partner)
}
\arguments{
\item{comm}{(required) metacommunity matrix indicating which 
association is present in microsite of each site. Rows are sites
and columns are microsites/individuals.}

\item{topo_names}{(required) matrix representing association network 
where integers label each association, as produced by 
\code{\link{name_topo}}. Labels should match those used in \code{comm}.}

\item{pool}{(required) binary site x species matrix indicating which species
 are present in each site's species pool, but which are not necessarily
established.}

\item{mortality}{(required) probability that an unassociated mutalist dies}

\item{partner}{(required) integer indicating whether the function
applies to hosts (\code{1}) or symbionts (\code{2}).}
}
\value{
binary site x species matrix indicating which species remain
	in the species' pools at each site
}
\description{
Random mortality of species in communities that are not
established in a microsite as a mutualistic association.
}

