% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disperse.r
\name{disperse}
\alias{disperse}
\title{Dispersal}
\usage{
disperse(sites, niches, pool, gsad)
}
\arguments{
\item{sites}{(required) a 2 column matrix giving environmental conditions
at each site}

\item{niches}{(required) array of niche optima and breadths for each species
and each environmental axis (\code{[species, optima/breadth, environment]})
as returned by \code{\link{make_niches_gsad}}.}

\item{pool}{(required) binary site x species matrix indicating which species
 are present in each site's species pool, but which are not necessarily
established.}

\item{gsad}{vector of integers giving the number of propagules for each species
arriving at a site in a timestep}
}
\value{
binary site x species matrix indicating which species are now
	in the species' pools at each site
}
\description{
Stochastic dispersal of species into communities based on the 
global species abundance distribution
}
\details{
The global species abundance distribution \code{gsad} gives
the number of propagules of each species arriving at each site
in one time step. The probability that a species establishes 
is the probability that at least on propagule establishes, given
a per-propagule establishment probability calculated by 
\code{\link{niche_func}} based on the match between species niches
and environmental conditions in the site. The function assumes that 
hosts and symbionts disperse independently (no co-dispersal), since it
can only be applied to one or the other and not both concurrently.
}

