% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_camm.r
\name{initialize_camm}
\alias{initialize_camm}
\title{Initialize simulation}
\usage{
initialize_camm(parm_file = NA, save_start = F, runID = NA,
  save_dir = "./")
}
\arguments{
\item{parm_file}{path to file with parameters}

\item{save_start}{logical indicating whether metacommunity objects should be
saved to an .RData file. Defaults to \code{FALSE}.}

\item{runID}{character string identifying this simulation. Required if 
\code{save_start=TRUE}.}

\item{save_dir}{directory where simulation objects should be saved. Defaults
to current directory.}
}
\value{
a named list containing all objects needed to run a simulation
}
\description{
Initializes a simulation for a given set of parameters
}
\details{
The function creates host and symbiont species pools, global species 
abundance distributions (gsad), an association network (and named links),
an association probability matrix, a set of sites with environmental 
conditions, and an empty metacommunity at those sites. Parameters for
creating these objects are read from a file (\code{parm_file}) or from 
the environmentwhere the function is called. Required parameters are:
\describe{
\item{\code{S_a} and \code{S_b}}{number of host and symbiont species}
\item{\code{N_L}}{number of links in the association network} 
\item{\code{topology}}{association network topology (\code{'one2one'},
	\code{'one2many'} or \code{'many2many'}). See \code{\link{make_topo}}.}
\item{\code{N_C}}{number of sites}
\item{\code{N}}{number of microsites/individuals per site}
\item{\code{rho_z}}{correlation between environmental axes across sites}
\item{\code{niche_parms_a} and \code{niche_parms_b}}{named list of parameters
	used to create host (a) and symbiont (b) niches. 
	See \code{\link{make_niches_gsad}}.}
\item{\code{gsad_dist_a} and \code{gsad_dist_b}}{named list of parameters
	used to create host (a) and symbiont (b) global species abundance 
	distributions. See \code{\link{make_niches_gsad}}.}
}
Optional parameters are:
\describe{
\item{comm_fill}{character string indicating how initial metacommunity
	should be filled. Defaults to \code{'empty'}. See \code{link{make_comm}}.}
}
}
\seealso{
\code{link{make_topo}}, \code{link{make_niches_gsad}}, 
	\code{link{make_comm}}, \code{\link{make_sites}}
}

