% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_comm.r
\name{make_comm}
\alias{make_comm}
\title{Creation of metacommunity}
\usage{
make_comm(N_C, N, N_S = 0, fill = "empty")
}
\arguments{
\item{N_C}{(required) number of communities (sites)}

\item{N}{(required) number of microsites (or individuals) in each community}

\item{N_S}{number of species pairs}

\item{fill}{character string indicating how communities should be filled.
Defaults to \code{'empty'}. See details.}
}
\description{
Create a set of \code{N_C} communities, each with \code{N} microsites
or individuals.
}
\details{
A metacommunity is represented by a \code{N_C x N} matrix where rows give
the host-symbiont pair present at each site. Integers in the matrix
correspond to host-symbiont pairs named in the association network topology
matrix. Currently implements two methods for starting a metacommunity which
is specified by \code{fill}:
\describe{
\item{\code{'empty'}}{initial metacommunity has no species present}
\item{\code{'random'}}{initial metacommunity community populated by randomly 
	sampling species pairs until it is full}
}
}

