% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_niches_gsad.r
\name{make_niches_gsad}
\alias{make_niches_gsad}
\title{Creation of species niches and global abundances}
\usage{
make_niches_gsad(N_S, nicheparms, distribution)
}
\arguments{
\item{N_S}{(required) number of species}

\item{nicheparms}{(required) a named list decribing the distributions from 
    which niche optima and breadths are sampled. Contains:
\describe{
    \item{mu}{vector of length 2 with the maximum niche optima for each 
        environmental axis}
    \item{rho}{correlation between niche optima aross the two environmental
        axes}
    \item{sigma}{vector of length 2 with the means of the gamma 
        distributions from which niche breadths are sampled, corresponding
        to the two environmental axes}
    \item{alpha}{vector of length 2 with the shape parameters of the gamma
        distributions from which niche breadths are samples, corresponding
        to the two environmental axes}
    \item{r}{correlation between niche breadth across the two environmental
        axes}
}}

\item{distribution}{(required) named list of parameters describing the 
form of the species abundance distribution. Contains:
\describe{
    \item{type}{character string indicating the form of the distribution.
        See details.}
    \item{maxN}{maximum abundance. Used when other distribution parameters
        are not supplied.}
    \item{P_maxN}{probability of sampling an abundance greater than 
        \code{maxN}. Used in conjunction with \code{maxN} to define the
        shape of the distribution when other parameters are not supplied.}
    \item{p1}{parameter controling the shape of the distribution. 
        Interpretation depends on \code{type}. See details.}
    \item{p2}{parameter controling the shape of the distribution. 
        Interpretation depends on \code{type}. See details.}
    \item{corr}{matrix containing correlations between global abundance and
        niche means and breadths. See details.}
}}
}
\value{
named list whose first element is an array with species' niche 
	optima and breadths along each environmental axis and whose second element
	is a vector of species relative global abundances
}
\description{
Concurrently generates random 2-D gaussian niches and a species abundance
distribution that is correlated with niche parameters.
}
\details{
This function randomly generates niche optima and breadths for \code{N_S} species
along two envronmental axes as well as a global species abundance
distribution. Niche optima and breadths may be correlated with global 
abundance along one or both environmental axes. In additon, niche optima and 
breadths may also be correlated across axes. These correlations are
controlled by parameters in the named lists \code{nicheparmes} and 
\code{distribution} and are accomplished through Gaussian copulas. This method
samples five variables (niche optima on axes 1 and 2, niche breadths on axes
1 and 2, and species global abundance) from a multivariate normal 
distribution with covariance matrix defined by \code{r} (correlation between
niche optima across environmental axes), \code{rho} (correlation between 
niche breadths across environmental axes), and \code{corr} (correlation 
between global abundance and each niche parameter, given as a 2x2 matrix 
where columns refer to optima and breadths and rows refer to environmental 
axes). Then, niche optima are transformed to a uniform distribution on the  
interval \code{[-mu, mu]} for each environmental axis. Niche breadths
are transformed to a gamma distribution with mean \code{sigma} and shape 
parameter \code{alpha} for each environmental axis. Species abundances are 
transformed to the distribution given in \code{type}. The following forms
of global species abundance distributions are supported and are described 
along with thie requried parameters.
\describe{
\item{\code{'same'}}{All species have the same abundance.}
\item{\code{'uniform'}}{Uniform distribution on \code{[1, maxN]}.
	Can also specify endpoint with \code{p1}.}
\item{\code{'power'}}{Power-law distribution with power \code{p1}.
	Can alternatively specify \code{maxN} and \code{P_maxN} to find  
	\code{p1} corresponding to a distribution with a probability of  
	\code{P_maxN} of exceeding \code{maxN}.}
\item{\code{'logseries'}}{Logseries distribution where \code{p1} is
	Fisher's alpha and \code{p2} is N. Can alternatively specify 
	\code{maxN} and \code{P_maxN} to find N corresponding to a distribution
	with a probability of \code{P_maxN} of exceeding \code{maxN}.}
\item{\code{'lognormal'}}{Lognormal distribution with mean \code{p1}
	and std dev \code{p2}. Can alternatively specify \code{maxN} and 
	\code{P_maxN} to find the std dev corresponding to a distribution with 
	mean  = 0 and a probability of code{P_maxN} of exceeding \code{maxN}.}
\item{\code{'poisson'}}{Poisson distribution with rate (lambda) = 
	\code{p1}. Can alternatively specify \code{maxN} and \code{P_maxN} to   
	find \code{p1} corresponding to a distribution with a probability of  
	\code{P_maxN} of exceeding \code{maxN}.}
}
Note that is \code{maxN} is specified, it will be used instead of 
\code{p1} or \code{p2}. In addition, continuous distributions are discretized
so that abundances can be used as measures of propagule supply.
}
\note{
Users should take care when specifying correlations so that the 
results covariance matrix for niche parameters and global abundance is
positive definite. For example, if \code{r = 0.5} and global abundance is 
correlated with niche optima along environmental axis 1, it must also be 
correlated with niche optima along axis 2.
}

