% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_parmlist.r
\name{make_parmlist}
\alias{make_parmlist}
\title{Create parameter list}
\usage{
make_parmlist(e = parent.frame())
}
\arguments{
\item{e}{environment where parameters are located}
}
\value{
a named list of paramters
}
\description{
Creates a list of simulation parameters
}
\details{
The function creastes a list of parameters required for running a 
simulation. Parameters are taken by default from the environment where
the function is called, or optionally from the environment provided
by \code{e}. All parameters are \strong{required} for simulation and \emph{must be present
in the environment or the function will fail}.

Parameters for creating a metacommunity (see \code{\link{initialize_camm}}):
\itemize{
\item \strong{\code{runID}} : name of this simulation run
\item \strong{\code{S_a}} : number of host species
\item \strong{\code{S_b}} : number of symbiont species
\item \strong{\code{N_C}} : number of sites
\item \strong{\code{N}} : number of microsites / individuals
\item \strong{\code{rho_z}} : correlation between environmental axes
	across sites 
\item \strong{\code{N_L}}: number of links in the association network 
	between host and symbiont species
\item \strong{\code{topology}}: character string defining type of 
	association network topology
\item \strong{\code{N_L}}: number of links in the association network 
	between host and symbiont species
}
Parameters for creating species niches (see \code{\link{make_niches_gsad}}):
\itemize{
\item \strong{\code{mu_a1}} and \strong{\code{mu_a2}} : maximum niche
	optima for host species along environmental axes 1 and 2
\item \strong{\code{mu_b1}} and \strong{\code{mu_b2}} : maximum niche
	optima for symbiont species along environmental axes 1 and 2
\item \strong{\code{rho_a}} and \strong{\code{rho_b}}: correlation of niche
	optima between environmental axes across host and symbiont species
\item \strong{\code{sigma_a1}} and \strong{\code{sigma_a2}} : expected
	niche breadth of host species along environmental axes 1 and 2
\item \strong{\code{sigma_b1}} and \strong{\code{sigma_b2}} : expected
	niche breadth of symbiont species along environmental axes 1 and 2
\item \strong{\code{alpha_a1}} and \strong{\code{alpha_a2}} : shape
	parameter of the gamma distribution from which host niche breadths
	are drawn for environmental axes 1 and 2
\item \strong{\code{alpha_b1}} and \strong{\code{alpha_b2}} : shape
	parameter of the gamma distribution from which host niche breadths
	are drawn for environmental axes 1 and 2
\item \strong{\code{r_a}} and \strong{\code{r_b}}: correlation of niche
	breadths between environmental axes across host and symbiont species
\item \strong{\code{gsad_dist_a}} and \strong{\code{gsad_dist_b}}: list
	of paramters describing the global species abundance distribution for
	hosts ans symbionts
}
Parameters for running the simulation:
\itemize{
\item \strong{\code{omega}}: degree of mutualism: \code{0} = no 
	facilitation, \code{1} = obligate mutualism 
	(see \code{\link{calc_probs}})
\item \strong{\code{assoc_prob_func}}: function that defines the
	probability that a host and symbiont will form an association, given
	that they have the capacity to. Should accept three arguments: 
	\code{env} (vector of environmental conditions), \code{niche_a} 
	(matrix defining the host's niche) and \code{niche_b} (matrix defining
	the symbiont's niche).
\item \strong{\code{mort_rate}}: probability that an established association
	dies in a timestep
\item \strong{\code{mort_rate_a}} and \strong{\code{mort_rate_b}}: local
	extinction rate of unassociated hosts and symbionts from a species pool
	relative to the mortality rate of established associations (e.g., a
	value of \code{10} will cause cause unassociated mutualists to be removed
	from a species pool at \code{10x} the rate of associated mutualists)
}
}

