% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sites.r
\name{make_sites}
\alias{make_sites}
\title{Creation of sites}
\usage{
make_sites(N_C, rho_z = 0)
}
\arguments{
\item{N_C}{(required) number of sites}

\item{rho_z}{correlation between two environmental conditions}
}
\value{
a matrix of values with dimension \code{N_C x 2}
}
\description{
Randomly generates sites with two potentially correlated environmental
conditions. Conditions are drawn from a multivariate normal distribution
with mean = 0 and variance = 1.
}

