% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_topo.r
\name{make_topo}
\alias{make_topo}
\title{Creation of an association network}
\usage{
make_topo(S_a, S_b, N_L, topology)
}
\arguments{
\item{S_a}{(required) number of host species}

\item{S_b}{(required) number of symbiont species}

\item{N_L}{(required) number of edges (links) in the network}

\item{topology}{(required) form of network topology. Can be 
'one2one', 'one2many', 'many2many' (see details).}
}
\value{
a binary matrix with dimension \code{S_a x S_b} indicating links
	between host and symbiont species
}
\description{
Creates an association network between a given number of host and symbiont
species with a defined network topology.
}
\details{
The function generates a random bipartite network between host and symbiont
species which is returned as a matrix of 0s and 1s (indicating whether two 
species can associate with one another). Users can specify the network topology
as:
\describe{
\item{\code{'one2one'}}{each host associates with one symbiont and each symbiont
	associates with one host}
\item{\code{'one2many'}}{each host associates with one symbiont and each symbiont
	associates with one or more hosts}
\item{\code{'many2many'}}{each host associates with one or more symbionts and
	each symbiont associates with one or more hosts}
}

The specified topology must be compatible with the number of species and 
number of links (\code{N_L}) specified. E.g. for \code{'one2one'}, the 
number of of host and symbiont species and links must be equal, whereas, for
\code{'one2many'} the number of links must be greater than the number of symbionts.
}

