% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/niche_func.r
\name{niche_func}
\alias{niche_func}
\title{Calculate survival probability}
\usage{
niche_func(x, mu, sigma)
}
\arguments{
\item{x}{(required) vector of environmental conditions}

\item{mu}{(required) vector of niche optima (matching x)}

\item{sigma}{(required) vector of niche breadths (matching x)}
}
\value{
the probability of establishment
}
\description{
Calculates the probability that a propagule will establish
under specific environmental conditions
}
\details{
The probability that a propagule will establish is the product
of the probabilities that the propagule will establish under each
environmental condition. (E.g. it assumes independent effects of the
environmental axes.) Probabilities are calculated using a Gaussian curve
whose mean is the niche optimum and whose standard deviation is the 
niche breadth. Curves are not standardized to be pdfs; thus, each curve
reaches a maximum of 1 when the environment exactly matches a species 
optimum.
}

