% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_camm.r
\name{run_camm}
\alias{run_camm}
\title{Run a simulation}
\usage{
run_camm(metacomm = NULL, reps = NA, save_steps = NA)
}
\arguments{
\item{metacomm}{list containing objects describing the metacommunity, 
as returned by \code{\link{initialize_camm}}. If not supplied, then then
environment must contain the following:
\describe{
    \item{\code{comm}}{metacommunity object (see \code{make_comm)})}
    \item{\code{poolA} and \code{poolB}}{species pools for hosts and
        symbionts (see \code{calc_pool})}
    \item{\code{niches_a}, \code{niches_b}, \code{gsad_a}, \code{gsad_b}}{
        host ans symbiont niche parameters and global species abundance 
        distributions (see \code{make_niches_gsad})}
    \item{\code{assoc_probs}}{site x association matrix giving the
        probability of each association forming at each site}
    \item{\code{topo} and \code{topo_names}}{binary host x symbiont species
        matrix giving the association network (see \code{make_topo}) and 
        a versionwith integers identifying each association (see 
        \code{name_topo})}
}}

\item{reps}{(required) number of timesteps to run simulation}

\item{save_steps}{vector indicating which timesteps to save. Defaults to all.}
}
\value{
a list containing objects describing states of the metacommunity 
	through time (see details)
}
\description{
Runs a simulation on an existing metacommunity
}
\details{
Simulations are run for a fixed number of timesteps (\code{reps}) on the 
metacommunity provided in \code{metacomm}, or if not provided, on the 
metacommunity objects that exist in the environment where the function is 
called. The simulation keeps track of the associations (\code{comm}),
host species (\code{poolA}), and symbiont species (\code{poolB}) present
at each site at each timestep to be saved (\code{save_steps}, defaults to all.
It also records the transition matrix used to generate the status of the 
at each of these timesteps (\code{Tmat}. These objects are returned as a 
list.

For each timestep, the simulation executes the following operations in 
order:
\enumerate{
	\item \strong{Local Extinction} : Stochastic extinction of species from 
		sites where they are not established in an association at a microsite. 
		See \code{\link{die}}.
	\item \strong{Immigration} : Stochastic immigration of species from mainland
		into species pools at local sites. The probability that a species becomes
		established in a site's local pool is a function of the number of propagules
		arriving and the match between the species' niche and environmental
		conditions. See \code{\link{disperse}}.
	\item \strong{Transition Communities} : Calculate a transition matrix
		for each site based on species present in the species pool and then
		stochastically change microsite states based on these transition
		probabilities. This is the step where established associations can
		die and empty microsites are colonized from species present in local
		site species pools. See \code{\link{calc_probs}} and 
		\code{\link{transition}}.
}
}
\note{
A previous version of the function allowed the simulation to run until
multiple 'chains' reached a convergence criteria. This mode has been removed,
but the code remains commented out in the function in case a future user wants
to utilize it.
}

