% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_camm_N.r
\name{run_camm_N}
\alias{run_camm_N}
\title{Run multiple simulations}
\usage{
run_camm_N(parm_file, nruns, nchains, nparallel = 1, sim_parms, simID,
  save_start = F, save_sim = F, save_dir = "./", restart = F,
  sim_dir = NULL, dev = FALSE)
}
\arguments{
\item{parm_file}{(required) path to file where parameters are stored. 
See \code{\link{write_parms}} and \code{\link{make_parmlist}} for 
instructions on creating an appropriate parameter file}

\item{nruns}{(required) number of independent simulation runs to conduct}

\item{nchains}{(required) number of replicate simulations on each initial
metacommunity. Defaults to 1.}

\item{nparallel}{number of cores used for parallel simulation. Defaults to
1.}

\item{sim_parms}{(required) named list of parameters controlling simulation.
Should contain:
\describe{
    \item{reps}{vector of timepoints at which to summarize simulations
        across chains (see details). The last number is the timepoint at
        which simulations are stopped.}
    \item{save_steps}{optional vector of timepoints to save. Defaults to 
        all.}
}}

\item{simID}{(required) character vector identifying this simulation}

\item{save_start}{logical indicating whether the initial metacommunities
should be saved to an RData file. Defaults to \code{FALSE}.}

\item{save_sim}{logical indicating whether simulations results should
be save to RData files. Defaults to \code{FALSE}.}

\item{save_dir}{path to directory where simulation results should be saved.
Defaults to the current directory.}

\item{restart}{logical indicating whether this simulation restarts an existing
simulation. If \code{FALSE}, existing files will be overwritten. If 
\code{TRUE}, then a metacommunity object named 
\code{'<runID>_metacomms.RData'} must be present in \code{save_dir}.
Defaults to \code{FALSE}.}

\item{sim_dir}{directory where CAMM is installed. Defaults to R's search path.}

\item{dev}{logical flag indicating whether the function is being called in
development mode, in which case the package is loaded using devtools from
\code{sim_dir}}
}
\value{
a list of lists (one for each independent simulation run), each 
	containing two arrays summarizing diversity ('rich_stats') and 
	community-environment correlations ('corr_stats') across simulations run   
	on the same initial metacommunity (across chains). The first dimension of 
	these arrrays refers to whether the quantities are the cross-chain mean
	or variance. If summaries were requested for multiple timepoints, then 
	the timepoint is indicated in the last dimension.
}
\description{
Runs multiple simulations on the same set of parameters.
}
\details{
This function runs multiple simulations on a set of parameters 
given in \code{parm_file}. The function first generates \code{nruns}
independent metacommunities and if \code{save_start} is \code{TRUE}, 
saves these as a list in the directory given in \code{save_dir} as
\code{'<simID>_metacomms.RData'}. The function then executes one or more
simulations (given in \code{nchains}) on each of these metacommunities.
This allows the user to evaluate variability among simulations with 
different initial metacommunities versus variability that arises during 
the simulation. Simulations are controled by parameters passed in 
\code{sim_parms}, including the number of timesteps and which steps to 
record. Mean diversity statistics across communities ('rich_stats') and  
community-environment correlations ('corr_stats') are calculated for  
each simulation using \code{\link{calc_commstats}} and 
\code{\link{calc_envcorr}}) and the mean and variance of these community 
descriptors are calculated across simulations run on the same initial  
metacommunity (across 'chains'). Summaries can be calculated for multiple  
timepoints by passing a vector of timepoints to the \code{sim_parms$reps}.   
In this case, the simulation will be run until the last timepoint in the  
vector (not the largest). If \code{save_sim} is \code{TRUE} then both the  
cross-chain summary statistics and full simulation results are saved to 
files in \code{save_dir} as \code{'<simID>_<run#>_results.RData'} and 
\code{'<simID>_<run#>_metacomms-end.RData'}, respectively. Otherwise
only the summaries are returned by the function, in a list.

Simulations can be run in parallel by specifying 
\code{nparallel > 1}, which requires the \code{\link[doParallel]{doParallel}}
and \code{\link[foreach]{foreach}} packages.
By default, \code{nparallel = 1} and the simulations proceed serially.
}

