% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition.r
\name{transition}
\alias{transition}
\title{Transition microsite}
\usage{
transition(probs)
}
\arguments{
\item{probs}{(required) named vector of probabilities that a microsite will
take on a given value in the next timestep. Names are the state values.}
}
\value{
an integer representing the new state
}
\description{
Stochastically transitions a microsite in a community based on transition
probabilities.
}

