% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_parms.r
\name{write_parms}
\alias{write_parms}
\title{Write parameter file}
\usage{
write_parms(parm_list, file_name, file_dir = "./")
}
\arguments{
\item{parm_list}{(required) named list of parameters to record}

\item{file_name}{(required) path to file where parameters should be written}

\item{file_dir}{directory where file should be saved, if not included in 
\code{file_name}. Defaults to current directory.}
}
\value{
nothing
}
\description{
Write simulation parameters to a file
}
\details{
This function writes paramater assignments from a named list of parameters
to a given file. Each element in the list is written as a line that
can be read into R (using \code{\link{source}}) such the name of the list
element becomes the object name. The function only works for vectors, lists
with one level, and functions. It then appends two lines defining 
niche parameter lists for hosts and symbionts.
}

